+++ .env
@@ -36,7 +36,7 @@ OAUTH_GITHUB_CLIENT_SECRET=10a7cb9c813e5c737bd8783bd89aacad9a479aaa
 # For an SQLite database, use: "sqlite:///%kernel.project_dir%/var/data.db"
 # For a PostgreSQL database, use: "postgresql://db_user:db_password@127.0.0.1:5432/db_name?serverVersion=11&charset=utf8"
 # IMPORTANT: You MUST configure your server version, either here or in config/packages/doctrine.yaml
-DATABASE_URL=mysql://developer:qwerty123@127.0.0.1:3306/symfony?serverVersion=5.7
+DATABASE_URL=mysql://developer:qwerty123@127.0.0.1:3306/symfony2?serverVersion=5.7
 ###< doctrine/doctrine-bundle ###
 
 ###> symfony/swiftmailer-bundle ###
