+++ config/services.yaml
@@ -24,8 +24,7 @@ services:
     App\Controller\:
         resource: '../src/Controller'
         tags: ['controller.service_arguments']
-
-   
+       
     #App\DataFixtures\AppFixtures:
         #$slugify: 'Cocur\Slugify\Slugify'
