--- src/Repository/UserNetRepository.php
@@ -1,58 +1,0 @@
-<?php
-
-namespace App\Repository;
-
-use App\Entity\UserNet;
-use Doctrine\Bundle\DoctrineBundle\Repository\ServiceEntityRepository;
-use Doctrine\Persistence\ManagerRegistry;
-
-/**
- * @method UserNet|null find($id, $lockMode = null, $lockVersion = null)
- * @method UserNet|null findOneBy(array $criteria, array $orderBy = null)
- * @method UserNet[]    findAll()
- * @method UserNet[]    findBy(array $criteria, array $orderBy = null, $limit = null, $offset = null)
- */
-class UserNetRepository extends ServiceEntityRepository
-{
-    public function __construct(ManagerRegistry $registry)
-    {
-        parent::__construct($registry, UserNet::class);
-    }
-
-    // /**
-    //  * @return UserNet[] Returns an array of UserNet objects
-    //  */
-    /*
-    public function findByExampleField($value)
-    {
-        return $this->createQueryBuilder('u')
-            ->andWhere('u.exampleField = :val')
-            ->setParameter('val', $value)
-            ->orderBy('u.id', 'ASC')
-            ->setMaxResults(10)
-            ->getQuery()
-            ->getResult()
-        ;
-    }
-    */
-
-    /*
-    public function findOneBySomeField($value): ?UserNet
-    {
-        return $this->createQueryBuilder('u')
-            ->andWhere('u.exampleField = :val')
-            ->setParameter('val', $value)
-            ->getQuery()
-            ->getOneOrNullResult()
-        ;
-    }
-    */
-    public function loadUserByUsername(string $email)
-    {
-        return $this->createQueryBuilder('u')
-            ->where('u.email = :email')
-            ->setParameter('email', $email)
-            ->getQuery()
-            ->getOneOrNullResult();
-    }
-}
